remove useEffect from 'useSubmitProposal' into own hook.
Extracts `CanUserSumitProposal` logic to its own hook

- This was adding a few extra request through out the app where there wasn't a need to call this again.
